refactor: Rename exception hanlder packages and clases
- errorhandler to be exceptionhandler
- Error*Handler classes to be Exception*Handler classes
